feat: Add right click on page to enable
This is still a grand mess. Right now, you can right click on a web
page, go to the context menu item for the app, and enable Auto Reload.
It's only going to work once in it's current state, meaning there is no
persistence to the auto reload. It's done once and never again. Also,
just to make sure I wasn't going crazy, it's every five seconds right
now for testing purposes. Like I said in the README, don't use this
app... yet.